Add Redis and MongoDB client classes
This commit adds the RedisClient and DBClient classes to the storage module. The RedisClient class is responsible for creating a Redis client and handling the connection to the Redis server. The DBClient class is responsible for creating a MongoDB client and handling the connection to the MongoDB server.

These classes will be used for caching and database operations in the webmart project.