feat: Add VerifyToken middleware to secure authentication routes
This commit adds the VerifyToken middleware to the authentication routes in the AuthRoute.js file and UserRoute.js file. The VerifyToken middleware is responsible for verifying the authenticity of the JSON web token (JWT) included in the request headers. By adding this middleware, we ensure that only authenticated users can access the protected routes for changing passwords and logging out. This change enhances the security of the webmart project.