feat: Add react-toastify npm dependency for displaying notifications
This commit adds the react-toastify npm dependency to the frontend package.json file. The react-toastify library is used for displaying notifications to the user. By including this dependency, we can easily show informative and user-friendly notifications throughout the application. This change enhances the user experience and provides important feedback to the user.